Decompose the routes a little to make it clearer to read
