+++ enclave-manager/web/src/emui/enclaves/enclave/Enclave.tsx
@@ -1,57 +1,13 @@
 import { Button, Flex, Spinner, Tab, TabList, TabPanel, TabPanels, Tabs } from "@chakra-ui/react";
 import { FiEdit2 } from "react-icons/fi";
-import { Await, LoaderFunctionArgs, useParams, useRouteLoaderData } from "react-router-dom";
-import { KurtosisClient } from "../../client/enclaveManager/KurtosisClient";
-import { EnclaveOverview } from "../../components/enclaves/EnclaveOverview";
+import { Await, useParams, useRouteLoaderData } from "react-router-dom";
+import { EnclaveOverview } from "../../../components/enclaves/EnclaveOverview";
 
 import { Suspense } from "react";
-import { DeleteEnclavesButton } from "../../components/enclaves/widgets/DeleteEnclavesButton";
-import { KurtosisAlert } from "../../components/KurtosisAlert";
-import { isDefined } from "../../utils";
-import { EnclavesLoaderResolved } from "./EnclaveList";
-
-export const enclaveLoader =
-  (kurtosisClient: KurtosisClient) =>
-  async ({ params }: LoaderFunctionArgs): Promise<{ routeName: string }> => {
-    const uuid = params.enclaveUUID;
-
-    if (!isDefined(uuid)) {
-      return {
-        routeName: "Missing uuid",
-      };
-    }
-
-    const enclavesResult = await kurtosisClient.getEnclaves();
-    if (enclavesResult.isErr) {
-      return {
-        routeName: uuid,
-      };
-    }
-
-    const enclave = Object.values(enclavesResult.value.enclaveInfo).find((enclave) => enclave.shortenedUuid === uuid);
-    if (!isDefined(enclave)) {
-      return {
-        routeName: uuid,
-      };
-    }
-
-    return {
-      routeName: enclave.name,
-    };
-  };
-
-export const enclaveTabLoader = async ({ params }: LoaderFunctionArgs): Promise<{ routeName: string }> => {
-  const activeTab = params.activeTab;
-
-  switch (activeTab?.toLowerCase()) {
-    case "overview":
-      return { routeName: "Overview" };
-    case "source":
-      return { routeName: "Source" };
-    default:
-      return { routeName: "Overview" };
-  }
-};
+import { DeleteEnclavesButton } from "../../../components/enclaves/widgets/DeleteEnclavesButton";
+import { KurtosisAlert } from "../../../components/KurtosisAlert";
+import { isDefined } from "../../../utils";
+import { EnclavesLoaderResolved } from "../loader";
 
 export const Enclave = () => {
   const { enclaves } = useRouteLoaderData("enclaves") as EnclavesLoaderResolved;
