+++ enclave-manager/web/src/emui/enclaves/enclave/tabLoader.ts
@@ -1,0 +1,14 @@
+import { LoaderFunctionArgs } from "react-router-dom";
+
+export const enclaveTabLoader = async ({ params }: LoaderFunctionArgs): Promise<{ routeName: string }> => {
+  const activeTab = params.activeTab;
+
+  switch (activeTab?.toLowerCase()) {
+    case "overview":
+      return { routeName: "Overview" };
+    case "source":
+      return { routeName: "Source" };
+    default:
+      return { routeName: "Overview" };
+  }
+};
